Last changes regarding refactoring, and cover delete languages

+++ src/Application/JobOffer/Commands/CloseJobs.cs
@@ -77,7 +77,9 @@ namespace Application.JobOffer.Commands
                 }
                 else
                 {
-                    _jobVacancyLanguageRepo.Delete(job.IdjobVacancy);
+                    var langs = _jobVacancyLanguageRepo.Get(job.IdjobVacancy);
+                    if (langs != null && langs.Any())
+                        _jobVacancyLanguageRepo.Delete(job.IdjobVacancy);
                     var ans = await _regJobVacWorkPermitRepo.Delete(job.IdjobVacancy);
                     var isPack = _contractProductRepo.IsPack(job.Idcontract);
                     if (isPack)
